+++ src/main/java/net/sf/mpxj/primavera/p3/P3DatabaseReader.java
@@ -160,7 +160,7 @@ public final class P3DatabaseReader extends AbstractProjectFileReader
          config.setAutoWBS(false);
 
          // Activity ID
-         m_projectFile.getCustomFields().getCustomField(TaskField.TEXT1).setAlias("Code");
+         m_projectFile.getCustomFields().getCustomField(TaskField.TEXT1).setAlias("Code").setUserDefined(false);
 
          m_projectFile.getProjectProperties().setFileApplication("P3");
          m_projectFile.getProjectProperties().setFileType("BTRIEVE");