+++ src/main/java/net/sf/mpxj/turboproject/TurboProjectReader.java
@@ -443,7 +443,7 @@ public final class TurboProjectReader extends AbstractProjectStreamReader
       CustomFieldContainer fields = m_projectFile.getCustomFields();
       for (Map.Entry<FieldType, String> entry : ALIASES.entrySet())
       {
-         fields.getCustomField(entry.getKey()).setAlias(entry.getValue());
+         fields.getCustomField(entry.getKey()).setAlias(entry.getValue()).setUserDefined(false);
       }
    }